backend_legacy/requirements.txt
fastapi
pydantic
uvicorn
ddgs
bs4
curl_cffi
google-genai
openai